fixed mermaid schema update after data task, moved refresh functions out of error.js
